requirements.txt
gradio>=4.44.0,<5.0.0
polars>=0.20.0
psycopg2-binary>=2.9.9
plotly>=5.20.0
python-dotenv>=1.0.0
